SlurmTools.jl

SlurmTools.jl is a very basic Julia interface to the Slurm cluster manager.

Core functions

SlurmTools.sacct — Function

sacct([options])

Gets accounting data for all jobs and job steps, returning a DataFrame.

Do not use any options that change the formatting of the output of sacct (e.g. --parsable, --helpformat, etc).

Example

julia> sacct(["--format=JobID,Start,End", "--allocations"])
340×3 DataFrames.DataFrame
│ Row │ JobID      │ Start               │ End                 │
│     │ String     │ DateTime            │ DateTime            │
├─────┼────────────┼─────────────────────┼─────────────────────┤
│ 1   │ 8646713    │ 2020-01-30T00:02:13 │ 2020-01-30T00:08:48 │
│ 2   │ 8646714    │ 2020-01-30T00:08:48 │ 2020-01-30T00:10:07 │
...

See also

sacct_formats

External links

sacct man page

SlurmTools.sbatch — Function

sbatch(cmd[, options])

Submits the script (and any arguments) specified by cmd as a batch job to Slurm, returning the job id. options is a collection of command-line options that might be passed

Example

julia> sbatch(`job.sh arg1`, ["--time=01:00:00", "--nprocs=4"])
"2332883"

External links

sbatch man page

SlurmTools.scancel — Function

scancel(job)

Cancel a queued Slurm job.

External links

scancel man page

Extra

SlurmTools.sacct_formats — Function

sacct_formats()

Returns a list of commands that can be passed with the "--format" option to sacct.